+++ llama_index/vector_stores/__init__.py
@@ -21,7 +21,7 @@ from llama_index.vector_stores.elasticsearch import (
 from llama_index.vector_stores.epsilla import EpsillaVectorStore
 from llama_index.vector_stores.faiss import FaissVectorStore
 from llama_index.vector_stores.lancedb import LanceDBVectorStore
-from llama_index.vector_stores.lantern import LanternStore
+from llama_index.vector_stores.lantern import LanternVectorStore
 from llama_index.vector_stores.metal import MetalVectorStore
 from llama_index.vector_stores.milvus import MilvusVectorStore
 from llama_index.vector_stores.myscale import MyScaleVectorStore
@@ -87,5 +87,5 @@ __all__ = [
     "TencentVectorDB",
     "AstraDBVectorStore",
     "AzureCosmosDBMongoDBVectorSearch",
-    "LanternStore",
+    "LanternVectorStore",
 ]